# public/openapi.yaml
openapi: 3.1.0
info:
  title: DomainDash API
  version: 1.0.0
  description: |
    DomainDash is a managed domain, hosting, SSL and internet-services portal backed by
    Synergy Wholesale, HaloPSA, IT Glue, IP2Whois and Vocus WSM.

    ## Authentication

    All API requests must include an `x-api-key` header containing a valid API key.
    Keys are created and managed by an Administrator in **Admin → API Keys**.

    Each key carries:
    - **Scopes** – `read` (GET endpoints) and/or `write` (mutation endpoints)
    - **IP allowlist** – comma-separated IPs or CIDR ranges; `*` allows any IP
    - **Rate limit** – configurable requests per hour (default 360)

    All requests are logged (method, path, IP, user-agent, response status).

    ## Base URL

    ```
    https://your-domaindash-instance.example/api
    ```

    ## Pagination

    List endpoints return a standard Laravel pagination envelope:

    ```json
    {
      "data": [...],
      "current_page": 1,
      "last_page": 3,
      "per_page": 100,
      "total": 250,
      "from": 1,
      "to": 100,
      "links": { "first": "...", "last": "...", "prev": null, "next": "..." }
    }
    ```

    ## Error responses

    | Status | Meaning |
    |--------|---------|
    | 401 | Missing or invalid API key |
    | 403 | IP not in allowlist, or scope not granted |
    | 404 | Resource not found |
    | 422 | Validation error |
    | 429 | Hourly rate limit exceeded |
    | 500 | Internal server error |

  contact:
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Obtain a key from Admin → API Keys. Supply it in the `x-api-key` request
        header. Keys must have the `read` scope for GET endpoints and the `write`
        scope for POST/PUT/PATCH/DELETE endpoints.

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    DnsRecord:
      type: object
      properties:
        id:
          type: integer
          example: 101
        domain_id:
          type: integer
          example: 42
        record_id:
          type: string
          nullable: true
          description: Synergy Wholesale DNS record ID
          example: "syn-dns-9988"
        host:
          type: string
          nullable: true
          description: Subdomain or blank for the zone apex (@)
          example: mail
        type:
          type: string
          enum: [A, AAAA, CNAME, MX, TXT, NS, SRV, CAA, PTR]
          example: MX
        content:
          type: string
          description: Record value (IP address, hostname, TXT string, etc.)
          example: mail.example.com.au
        ttl:
          type: integer
          description: Time-to-live in seconds (60–604800)
          default: 3600
          example: 3600
        prio:
          type: integer
          description: Priority — relevant for MX and SRV records
          default: 0
          example: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DnsRecordWrite:
      type: object
      required:
        - type
        - content
      properties:
        host:
          type: string
          nullable: true
          description: Leave blank or omit for the zone apex.
          example: mail
        type:
          type: string
          enum: [A, AAAA, CNAME, MX, TXT, NS, SRV, CAA, PTR]
          example: MX
        content:
          type: string
          example: mail.example.com.au
        ttl:
          type: integer
          minimum: 60
          maximum: 604800
          default: 3600
          example: 3600
        prio:
          type: integer
          minimum: 0
          maximum: 65535
          default: 0
          example: 10
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Resource not found.
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - "The email field is required."

  # -------------------------------------------------------------------------
  # Reusable parameters
  # -------------------------------------------------------------------------
  parameters:
...
    DomainId:
      name: domain
      in: path
      required: true
      schema:
        type: integer
      description: DomainDash domain ID.
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: API key required
    Forbidden:
      description: IP not in allowlist or required scope not granted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Scope denied
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    TooManyRequests:
      description: Hourly rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Rate limit exceeded

# ---------------------------------------------------------------------------
# Global security
# ---------------------------------------------------------------------------
security:
  - ApiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/domains/{domain}/dns/{recordId}:
    put:
      tags:
        - DNS Records
      operationId: updateDnsRecord
      summary: Update a DNS record
      description: |
        Replaces an existing DNS record (delete + re-add at Synergy).
        Requires `write` scope.
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - name: recordId
          in: path
          required: true
          schema:
            type: string
          description: Synergy DNS record ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DnsRecordWrite'
      responses:
        '200':
          description: DNS record updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DnsRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags:
        - DNS Records
      operationId: deleteDnsRecord
      summary: Delete a DNS record
      description: Removes a DNS record from the Synergy Wholesale zone. Requires `write` scope.
      parameters:
        - $ref: '#/components/parameters/DomainId'
        - name: recordId
          in: path
          required: true
          schema:
            type: string
          description: Synergy DNS record ID
      responses:
        '204':
          description: Record deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...